+++ tests/qcoreapplication.py
@@ -1,0 +1,5 @@
+from PySide6.QtCore import QCoreApplication
+
+s = 'abc'
+s = QCoreApplication.translate("GitFlowAdvanceIntBranch", u"hidden", None)
+s = QCoreApplication.translate("GitFlowAdvanceIntBranch", u"hidden", 'some help')